pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cosmoserver"
version = "0.1.0"
description = "Cosmo server leveraging LLM's to automate your home"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.45",
    "cosmocore",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.14",
    "mcp>=1.10.0",
    "pydantic>=2.11.7",
    "sqlalchemy>=2.0.0",
    "strands-agents>=0.1.9",
    "strands-agents-tools>=0.1.7",
    "tomlkit>=0.13.0",
]

[project.scripts]
cosmo-server = "cosmoserver.cli:cli_main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[dependency-groups]
dev = [
    "strands-agents-builder>=0.1.4",
    "pyright>=1.1.390",
    "ruff>=0.12.1",
]

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__", "**/.venv", "**/node_modules", "**/.git"]
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.12"
typeCheckingMode = "basic"

[tool.ruff]
target-version = "py312"
line-length = 90

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "B008",  # Do not perform function call in argument defaults - FastAPI Depends() is safe
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["cosmoserver"]

[tool.uv.sources]
cosmocore = { git = "https://github.com/marchese29/CosmoCore" }
